requirements.txt
matplotlib==3.10.8
numpy==1.24.1
Pillow==9.4.0
scipy==1.16.3
torch==2.2.0
torchmetrics==1.8.2
torchvision==0.17.0
tqdm==4.64.1
wandb==0.23.1
